requirements.txt
fastapi>=0.110,<1
uvicorn>=0.27,<1
python-dotenv>=1.0,<2
streamlit>=1.31,<2
pandas>=2.0,<3
numpy>=1.23,<3
scikit-learn>=1.3,<2
joblib>=1.3,<2
